Move API server to backend
