+++ DEV_ENV.md
@@ -27,7 +27,7 @@ The data portal dev environment is a set of containers defined in [docker-compos
 
 Both the data portal frontend and backend services will automatically reload when their source code is modified, but they won't automatically rebuild when their dependencies (such as npm or pip package lists) change.
 
-To update the dev env to reflect changes to [frontend/package.json](frontend/package.json) or [backend/corpora/api_server/requirements.txt](backend/corpora/api_server/requirements.txt), run `make local-sync`
+To update the dev env to reflect changes to [frontend/package.json](frontend/package.json) or [backend/api_server/requirements.txt](backend/api_server/requirements.txt), run `make local-sync`
 
 ### Update Dev Data
 
@@ -36,7 +36,7 @@ The dev environment is initialized with AWS Secrets/S3 data in the [scripts/setu
 ### Make targets for managing dev:
 
 | Command                                            | Description                                                                        | Notes                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
-|----------------------------------------------------| ---------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| -------------------------------------------------- | ---------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `make local-init`                                  | Launch a new local dev env and populate it with test data.                         |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | `make local-start`                                 | Start a local dev environment that's been stopped.                                 |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | `make local-stop`                                  | Stop the local dev environment.                                                    |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
