+++ Makefile
@@ -50,7 +50,7 @@ prod-performance-test:
 
 .PHONY: local-backend
 local-backend:
-	$(MAKE) local-server -C ./backend/corpora/api_server
+	$(MAKE) local-server -C ./backend/api_server
 
 .PHONY: smoke-test-prod-build
 smoke-test-prod-build:
