+++ backend/corpora/api/v1/collection_id/publish.py
@@ -4,7 +4,7 @@ from backend.common.corpora_orm import CollectionVisibility
 from backend.common.entities import Collection
 from backend.common.utils.http_exceptions import ConflictException
 
-from backend.corpora.api_server.db import dbconnect
+from backend.api_server.db import dbconnect
 from backend.common.utils.http_exceptions import ForbiddenHTTPException
 from ..authorization import owner_or_allowed
