+++ backend/corpora/api/v1/curation/collections/collection_id/actions.py
@@ -2,7 +2,7 @@ from flask import g, jsonify
 
 from backend.corpora.api.v1.collection import update_collection
 from ..common import EntityColumns
-from backend.corpora.api_server.db import dbconnect
+from backend.api_server.db import dbconnect
 from backend.common.corpora_orm import CollectionVisibility
 from backend.common.entities import Collection
 from backend.common.utils.http_exceptions import MethodNotAllowedException, NotFoundHTTPException
