+++ backend/corpora/api/v1/curation/collections/collection_id/revision.py
@@ -1,6 +1,6 @@
 from flask import make_response, jsonify
 
-from backend.corpora.api_server.db import dbconnect
+from backend.api_server.db import dbconnect
 from backend.corpora.api.v1.collection import post_collection_revision_common
