+++ requirements.txt
@@ -24,7 +24,7 @@ scanpy
 scipy==1.7.3
 tenacity
 tiledb==0.13.2  # Portal's tiledb version should always be the same or older than Explorer's
--r ./backend/corpora/api_server/requirements.txt
+-r ./backend/api_server/requirements.txt
 rsa>=4.7 # not directly required, pinned by Snyk to avoid a vulnerability
 parameterized
 python-json-logger
