+++ tests/unit/backend/gene_info/test_v1.py
@@ -4,7 +4,7 @@ import json
 
 from backend.gene_info.api import ncbi_provider
 from tests.unit.backend.corpora.fixtures.environment_setup import EnvironmentSetup
-from backend.corpora.api_server.app import app
+from backend.api_server.app import app
 import xml.etree.ElementTree as ET
 from unittest.mock import patch, call
