+++ tests/unit/backend/wmg/api/test_v1.py
@@ -2,7 +2,7 @@ import json
 import unittest
 from unittest.mock import patch
 
-from backend.corpora.api_server.app import app
+from backend.api_server.app import app
 from backend.wmg.api.v1 import find_dim_option_values
 from backend.wmg.data.schemas.cube_schema import cube_non_indexed_dims
 from tests.unit.backend.corpora.fixtures.environment_setup import EnvironmentSetup
